backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydicom==2.4.3
Pillow==10.1.0
numpy==1.26.2
requests==2.31.0
openai==1.3.5
python-dotenv==1.0.0
inference_sdk==0.9.0
google-generativeai
python-gdcm
pylibjpeg 